refactor: added useLinkSurveyUtils hook to abstract business logic
